elf: add rela from_fd method and implement for elf::from_path

+++ src/elf/dyn.rs
@@ -185,10 +185,10 @@ pub fn get_needed<'a, 'b>(dyns: &'a [Dyn], strtab: &'b Strtab<'a>, count: usize)
 
 /// Important dynamic LinkInfo generated via a single pass through the _DYNAMIC array
 pub struct LinkInfo {
-    pub rela: u64,
+    pub rela: usize,
     pub relasz: u64, // TODO: make this a usize?
     pub relaent: u64,
-    pub relacount: u64,
+    pub relacount: usize,
     pub gnu_hash: u64,
     pub hash: u64,
     pub strtab: usize,
@@ -246,10 +246,10 @@ impl LinkInfo {
         let mut soname = 0;
         for dyn in dynamic {
             match dyn.d_tag {
-                DT_RELA => rela = dyn.d_val.wrapping_add(bias), // .rela.dyn
+                DT_RELA => rela = dyn.d_val.wrapping_add(bias) as usize, // .rela.dyn
                 DT_RELASZ => relasz = dyn.d_val,
                 DT_RELAENT => relaent = dyn.d_val,
-                DT_RELACOUNT => relacount = dyn.d_val,
+                DT_RELACOUNT => relacount = dyn.d_val as usize,
                 DT_GNU_HASH => gnu_hash = dyn.d_val.wrapping_add(bias),
                 DT_HASH => hash = dyn.d_val.wrapping_add(bias),
                 DT_STRTAB => strtab = dyn.d_val.wrapping_add(bias) as usize,
